requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.1.1
cramjam==2.11.0
cycler==0.12.1
fastparquet==2024.2.0
fonttools==4.57.0
fsspec==2025.3.0
idna==3.11
importlib-resources==6.4.5
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.7.5
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==25.0
pandas==2.0.3
pillow==10.4.0
pyarrow==17.0.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.4
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.17.0
threadpoolctl==3.5.0
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.67.1
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
zipp==3.20.2
